탈퇴

+++ Project/src/main/java/egovframework/example/file/service/impl/FileServiceImpl.java
@@ -44,7 +44,7 @@ public class FileServiceImpl implements FileService {
     
     // 7/11 민중 게시글삭제를위한 달려있는 모든 파일 삭제 기능
     @Override
-    public void deleteAllByTargetIdAndType(int targetId, String useType) {
+    public void deleteAllByTargetIdAndType(Long targetId, String useType) {
         fileMapper.deleteByTargetIdAndType(targetId, useType);
     }
 }
